requirements.txt
customtkinter==5.0.1
Pillow==9.3.0
